팔로우, 생성, 수정, 삭제 Exception

+++ src/main/java/com/example/newsfeed/boards/service/BoardService.java
@@ -6,10 +6,16 @@ import com.example.newsfeed.boards.entity.Board;
 import com.example.newsfeed.boards.repository.BoardRepository;
 //import com.example.newsfeed.users.entity.User;
 //import com.example.newsfeed.users.repository.UserRepository;
+import com.example.newsfeed.exception.CustomException;
+import com.example.newsfeed.exception.ErrorCode;
+import com.example.newsfeed.user.entity.User;
+import com.example.newsfeed.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
+import java.awt.print.Pageable;
 import java.util.List;
 
 @Service
@@ -17,42 +23,51 @@ import java.util.List;
 public class BoardService {
 
     private final BoardRepository boardRepository;
-//    private final UserRepository userRepository;
-//
-//    public BoardResponseDto create(String contents, String username) {
-//
-//        User findUser = userRepository.findUserByUsernameOrElseThrow(username);
-//
-//        Board board = new Board(contents);
-//        board.setUser(findUser);
-//
-//        Board createBoard = boardRepository.save(board);
-//
-//        return new BoardResponseDto(createBoard.getId(), createBoard.getContents(), createBoard.getCreatedAt(), createBoard.getUpdatedAt());
-//    }
+    private final UserRepository userRepository;
+
+    public BoardResponseDto create(String contents, Long userId) {
+
+        User findUser = userRepository.findByIdOrElseThrow(userId);
+
+        Board board = new Board(contents);
+        board.setUser(findUser);
+
+        Board createBoard = boardRepository.save(board);
+
+        return new BoardResponseDto(createBoard.getId(), createBoard.getContents());
+    }
 
     public List<BoardResponseDto> findAll() {
         return boardRepository.findAll().stream().map(BoardResponseDto::toDto).toList();
     }
 
-//    public DetailBoardResponseDto findById(Long id) {
-//        Board findBoard = boardRepository.findByIdOrElseThrow(id);
-//        User writer = findBoard.getUser();
-//
-//        return new DetailBoardResponseDto(findBoard.getId(), findBoard.getContents(), writer.getUsername(), findBoard.getCreatedAt(), findBoard.getUpdatedAt());
-//    }
+    public DetailBoardResponseDto findById(Long id) {
+        Board findBoard = boardRepository.findByIdOrElseThrow(id);
+        User writer = findBoard.getUser();
+
+        return new DetailBoardResponseDto(findBoard.getId(), findBoard.getContents(), writer.getUsername(), findBoard.getCreatedAt(), findBoard.getUpdatedAt());
+    }
 
     @Transactional
-    public void updateBoard(Long id, String contents) {
+    public void updateBoard(Long id, String contents, Long userId) {
 
         Board findBoard = boardRepository.findByIdOrElseThrow(id);
+
+        if(!findBoard.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.BOARD_UPDATE_UNAUTHORIZED);
+        }
 
         findBoard.updateBoard(id, contents);
 
     }
 
-    public void deleteBoard(Long id) {
+    public void deleteBoard(Long id, Long userId) {
         Board findBoard = boardRepository.findByIdOrElseThrow(id);
+
+        if(!findBoard.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.BOARD_DELETE_UNAUTHORIZED);
+        }
+
         boardRepository.delete(findBoard);
     }
